datasets/census_bureau_international/_images/run_csv_transform_kub/requirements.txt
numpy
pandas
google-cloud-storage
gsutil
